+++ src/App.jsx
@@ -5,7 +5,7 @@ import Home from './routes/Home';
 import User from './routes/User';
 
 import './App.css';
-import UserSignUp from './components/UserSignUp.jsx';
+import Create from './routes/Create.jsx';
 
 function App() {
   const onAddEntry = async (communityId, selected) => {
@@ -19,7 +19,7 @@ function App() {
           <Route index path="/" element={<Home />} />
           <Route path="/add" element={<Add />} />
           <Route path="/user" element={<User />} />
-          <Route path="/user/create" element={<UserSignUp />} />
+          <Route path="/user/create" element={<Create />} />
         </Route>
         <Route path="*" element={<div>404: Error</div>} />
       </Routes>
